requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.30.0
stripe>=12.0.0
requests==2.31.0
beautifulsoup4==4.12.2
fpdf2==2.7.6
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
playwright==1.40.0
Jinja2==3.1.2
sendgrid==6.10.0 